loire: init: Force restorecon for /persist files
The restorecon_recursive directive in init is only
applied if the file_contexts file changed between
builds, but not necessarily if any file or folder
inside /persist has changed.

The restorecon code checks whether an xattr named
"security.restorecon_last" contains a string that
matches the current combined hashes of the SELinux
context files and skips restoring labels if there is
a match, see:
https://android.googlesource.com/platform/external/selinux/+/refs/tags/android-9.0.0_r35/libselinux/src/android/android_platform.c#1546

Force wiping that xattr so that restorecon always runs
since it's not very expensive (there are currently only
about 50 files on /persist).

The restorecon is needed to fix issues such as wrong
stock labels on:
/persist/hlos_rfs/: rfs_shared_hlos_file -> persist_rfs_file
/persist/rfs/: rfs_file -> persist_rfs_file
/persist/sensors/: sensors_persist_file -> persist_sensors_file

Change-Id: Icc9cf7be5dfbd8eaa783d073698cb630ddde7ba5