msm8956-common: hardware: lights: Adjust slightly off brightness conversion for maximum accuracy.
This change fixes a slight mistake that was introduced when trying to
improve accuracy.
https://github.com/sonyxperiadev/device-sony-common/commit/6199c66e80d7538c8f6633b742d83679e6bbf2a4
That patch aims to ensure maximum brightness is reached - which indeed
happens for value 0xff - but adds a small discrepancy as outlined below.

Consider the following example:

```python3
bits = 10
sbits = bits - 8
maxval = (1 << bits) - 1

for brightness in range(256):
    print(f'{brightness} => '
          f'original: {brightness << sbits | brightness >> sbits}, '
          f'proper: {round(brightness * maxval / 255)}, '
          f'fixed: {brightness << sbits | brightness >> (8 - sbits)}')
```

When running the example (see output below) a consistent (yet small)
bump can be observed because the highest bits are replicated in the
lower bits of the left-shifted value. This results an overlap between
the lowest and highest bits instead of filling just the zeroes that were
introduced by the left shift.
This also means that going over the limit of 4 extra bits (so 12 bits in
total) results in a too large right shift, making it unable to reach
maximum brightness.
The "fixed" value in the rightmost column solves both issues by making
sure that only the most significant bits end up in the lowest bits.

However, this only works for values up to 16 bits. While it is not
likely to see anything higher than 12, better be safe than
sorry. Adjusting brightness is not a hot code path *by any means* so we
can justify an integer division ;)

Sample output:

224 => original: 952, proper: 899, fixed: 899
225 => original: 956, proper: 903, fixed: 903
226 => original: 952, proper: 907, fixed: 907
227 => original: 956, proper: 911, fixed: 911
228 => original: 953, proper: 915, fixed: 915
229 => original: 957, proper: 919, fixed: 919
230 => original: 953, proper: 923, fixed: 923
231 => original: 957, proper: 927, fixed: 927
232 => original: 954, proper: 931, fixed: 931
233 => original: 958, proper: 935, fixed: 935
234 => original: 954, proper: 939, fixed: 939
235 => original: 958, proper: 943, fixed: 943
236 => original: 955, proper: 947, fixed: 947
237 => original: 959, proper: 951, fixed: 951
238 => original: 955, proper: 955, fixed: 955
239 => original: 959, proper: 959, fixed: 959

Signed-off-by: MarijnS95 <marijns95@gmail.com>
Change-Id: Ic347e5ff08226f6afec71c19fe41e099e8726c3d