msm8956-common: healthd: add support for storing battery parameters
There is a requirement to store battery parameters when the
system is running so that they may be restored when the system
boots up especially when the battery is removed and re-inserted.
Add support for it.

While at it, cleanup the code to reduce some redundancy. Also,
add a separate tag (healthd_msm) for printing logs from
libhealthd.msm during android mode.

CRs-Fixed: 1018233
Change-Id: I15439282c133ac92568bc7ff7750ab1b268cbb02