msm8956-common: power: Implement sustained performance power hint handling
Since version 7.0, Android supports a sustained performance
power hint for long-running applications, like games,
camera (in our case, viewfinder, photos but video encoding
keeps its own hint), RenderScript and audio processing:
in these cases the device requires to keep an high performance
state which will produce a lot of heat and, after a while,
the performance will decrease due to thermal mitigations.

To address this issue, handle the sustained performance hint
to limit the performance of the device to an acceptable level
while avoiding overheat.

Currently, we handle it only by limiting cores with rqbalance
but, in the future, it's planned to fine-tune the profile
with CPU frequency limits on cpufreq.

Change-Id: I6eaaa7d0f600a88ae1de60a0f86da7433302df09