msm8956-common: power: Implement Cluster Frequency Limits on hint functionality
This implements userspace management for the new RQBalance kernel
driver's Cluster Frequency Limits functionality.

Now we support setting limits on minimum and/or maximum frequency
for all clusters by specifying that on any RQBalance-PowerHAL
profile through the XML with the parameters:
- clusterN_freq_min to limit the minimum frequency
- clusterN_freq_max to limit the maximum frequency
where N=0 stands for the "first" cluster, typically LITTLE.

Even though this is not enforced, please note that this feature
is meant to be used with a voting mechanism for temporarily
limiting one or more clusters' frequency and this shall never
be used as a way to permanently set a minimum or maximum
frequency for the clusters: for this, please refer to
the standard CPUFreq sysfs nodes.

Change-Id: I13aec9d1f86fbd9f82b920e2668d18885ff3eb89