msm8956-common: TransPowerBase: Fix build
Source of TransPowerBase app is actually the same as
the source of TransPowerCommon library, so in Android O
it was enough to have an empty source folder and link
TransPowerCommon library as a static Android library.

However:
  1. the folder was missing in release
  2. this trick doesn't work in pre-O releases

The solution is to refer to appropriate source folder residing
in the common folder and also link to appropriate libraries.

Issue: https://github.com/sonyxperiadev/transpower/issues/4
Change-Id: Ia687bb96b0d8e3fdba4e0e0987ef62d2ab479895