msm8956-common: macaddrsetup: Create bluetooth file with 640 perms
The bluetooth service runs as bluetooth:bluetooth but the files created
by macaddrsetup are 600 system:system by default, so the bluetooth MAC
can't be read from /data/misc/bluetooth/bluetooth_bdaddr. This changes
the file's permissions to 640; a corresponding rc file change will put
the file in the bluetooth group.

Change-Id: I59b6ec7249de7795586d6f8c2a521162f8756a97