msm8956-common: hardware: transpower: libpower/transmitter: Don't consider SET_TRANSMIT_POWER failure fatal
We already know that SAR backoff doesn't work, and were previously
conveniently ignoring any failures.  The transition to TelephonyManager
in [#13] changed that and considers this to be fatal, but unless
external help comes in to guide testing on the new commands and command
structure (see open PRs) we should stick to the ignorant world and let
this misbehaviour slip through.  Which is "better" than crashing
transpower, in turn taking the telephony stack down and allowing
PartyCrasher (RescueParty, for the uninitiated) to make your phone
useless.

[#13]: https://github.com/sonyxperiadev/transpower/pull/13

Fixes: 21cdf15 ("libpower: Switch to TelephonyManager.")
Change-Id: I71fb2d1593838a64e9302c496d2b727cb6de94ba